deploy/requirements.txt
python-digitalocean==1.14.0